Break up state into separate keys when desired (#18)
We want to write to the disk as little as possible, but we write out the whole state any time any part of it changes. This change allows for the state to be optionally broken up into individual keys.

It does this by allowing serialize/1 callbacks to return a potentially nested map of binaries to be serialized separately by rocksdb. This allows us to not rewrite portions of the state that seldom change. This adds complexity on the behavior side, but it's worth it in certain cases for the reduced write throughput.